fix: run nx migrations
