+++ apps/react-demo/.babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["@nrwl/next/babel"],
+  "presets": ["@nx/next/babel"],
   "plugins": ["transform-typescript-metadata", "lodash"]
 }
