+++ apps/react-demo/project.json
@@ -5,10 +5,9 @@
   "projectType": "application",
   "targets": {
     "build": {
-      "executor": "@nrwl/next:build",
+      "executor": "@nx/next:build",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "root": "apps/react-demo",
         "outputPath": "dist/apps/react-demo"
       },
       "configurations": {
@@ -20,7 +19,7 @@
       "defaultConfiguration": "production"
     },
     "serve": {
-      "executor": "@nrwl/next:server",
+      "executor": "@nx/next:server",
       "options": {
         "buildTarget": "react-demo:build",
         "dev": true
@@ -38,13 +37,13 @@
       "defaultConfiguration": "development"
     },
     "export": {
-      "executor": "@nrwl/next:export",
+      "executor": "@nx/next:export",
       "options": {
         "buildTarget": "react-demo:build:production"
       }
     },
     "test": {
-      "executor": "@nrwl/jest:jest",
+      "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/apps/react-demo"],
       "options": {
         "jestConfig": "apps/react-demo/jest.config.ts",
@@ -52,7 +51,7 @@
       }
     },
     "lint": {
-      "executor": "@nrwl/linter:eslint",
+      "executor": "@nx/linter:eslint",
       "options": {
         "lintFilePatterns": ["apps/react-demo/**/*.{ts,tsx,js,jsx}"]
       },
