+++ jest.config.ts
@@ -1,3 +1,3 @@
-const { getJestProjects } = require('@nrwl/jest');
+const { getJestProjects } = require('@nx/jest');
 
 export default { projects: getJestProjects() };
