+++ jest.preset.js
@@ -1,3 +1,3 @@
-const nxPreset = require('@nrwl/jest/preset').default;
+const nxPreset = require('@nx/jest/preset').default;
 
 module.exports = { ...nxPreset };
