+++ libs/di/.eslintrc.json
@@ -5,7 +5,7 @@
     {
       "files": ["*.spec.ts", "*.test-d.ts"],
       "rules": {
-        "@nrwl/nx/enforce-module-boundaries": "off"
+        "@nx/enforce-module-boundaries": "off"
       }
     },
     {
