+++ libs/react-di/project.json
@@ -12,7 +12,7 @@
         "tsConfig": "libs/react-di/tsconfig.lib.json",
         "project": "libs/react-di/package.json",
         "entryFile": "libs/react-di/src/index.ts",
-        "rollupConfig": "@nrwl/react/plugins/bundle-rollup",
+        "rollupConfig": "@nx/react/plugins/bundle-rollup",
         "updateBuildableProjectDepsInPackageJson": false,
         "assets": [
           {
@@ -29,14 +29,14 @@
       }
     },
     "lint": {
-      "executor": "@nrwl/linter:eslint",
+      "executor": "@nx/linter:eslint",
       "options": {
         "lintFilePatterns": ["libs/react-di/**/*.{ts,tsx,js,jsx}"]
       },
       "outputs": ["{options.outputFile}"]
     },
     "test": {
-      "executor": "@nrwl/jest:jest",
+      "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/libs/react-di"],
       "options": {
         "jestConfig": "libs/react-di/jest.config.ts",
