+++ libs/react-utils/project.json
@@ -13,7 +13,7 @@
         "project": "libs/react-utils/package.json",
         "entryFile": "libs/react-utils/src/index.ts",
         "updateBuildableProjectDepsInPackageJson": false,
-        "rollupConfig": "@nrwl/react/plugins/bundle-rollup",
+        "rollupConfig": "@nx/react/plugins/bundle-rollup",
         "assets": [
           {
             "glob": "libs/react-utils/README.md",
@@ -29,14 +29,14 @@
       }
     },
     "lint": {
-      "executor": "@nrwl/linter:eslint",
+      "executor": "@nx/linter:eslint",
       "options": {
         "lintFilePatterns": ["libs/react-utils/**/*.{ts,tsx,js,jsx}"]
       },
       "outputs": ["{options.outputFile}"]
     },
     "test": {
-      "executor": "@nrwl/jest:jest",
+      "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/libs/react-utils"],
       "options": {
         "jestConfig": "libs/react-utils/jest.config.ts",
