+++ package.json
@@ -40,6 +40,7 @@
     "@faker-js/faker": "^7.6.0",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
+    "@swc/helpers": "~0.5.0",
     "core-js": "^3.22.7",
     "document-register-element": "1.14.10",
     "lodash": "^4.17.20",
@@ -59,22 +60,28 @@
     "@actions/github": "^5.1.1",
     "@babel/core": "7.12.13",
     "@babel/preset-env": "7.12.13",
-    "@babel/preset-react": "7.12.13",
+    "@babel/preset-react": "^7.14.5",
     "@babel/preset-typescript": "7.12.13",
     "@commitlint/cli": "^17.2.0",
     "@commitlint/config-conventional": "^17.2.0",
-    "@nrwl/cli": "15.0.13",
-    "@nrwl/devkit": "16.6.0",
-    "@nrwl/eslint-plugin-nx": "16.6.0",
-    "@nrwl/jest": "16.6.0",
     "@nrwl/js": "16.6.0",
-    "@nrwl/linter": "16.6.0",
-    "@nrwl/next": "16.6.0",
-    "@nrwl/nx-cloud": "16.2.0",
-    "@nrwl/react": "16.6.0",
-    "@nrwl/web": "16.6.0",
-    "@nrwl/workspace": "16.6.0",
+    "@nrwl/rollup": "16.6.0",
+    "@nrwl/webpack": "16.6.0",
     "@ns3/nx-npm": "^3.3.0",
+    "@nx/devkit": "16.6.0",
+    "@nx/eslint-plugin": "16.6.0",
+    "@nx/jest": "16.6.0",
+    "@nx/js": "16.6.0",
+    "@nx/linter": "16.6.0",
+    "@nx/plugin": "16.6.0",
+    "@nx/react": "16.6.0",
+    "@nx/web": "16.6.0",
+    "@nx/workspace": "16.6.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@svgr/webpack": "^6.1.2",
+    "@swc-node/register": "~1.4.2",
+    "@swc/cli": "~0.1.62",
+    "@swc/core": "~1.3.51",
     "@testing-library/react": "14.0.0",
     "@types/jest": "29.4.4",
     "@types/lodash": "^4.14.188",
@@ -87,6 +94,7 @@
     "babel-jest": "29.4.3",
     "babel-plugin-lodash": "^3.3.4",
     "commitizen": "^4.2.5",
+    "css-loader": "^6.4.0",
     "cz-conventional-changelog": "^3.3.0",
     "dotenv": "16.0.3",
     "eslint": "8.22.0",
@@ -101,15 +109,24 @@
     "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^8.0.2",
     "jest": "29.4.3",
+    "jest-environment-jsdom": "^29.4.1",
     "lint-staged": "^13.0.4",
     "nx": "16.6.0",
+    "nx-cloud": "16.2.0",
     "prettier": "2.7.1",
-    "react-test-renderer": "18.2.0",
+    "react-refresh": "^0.10.0",
     "semantic-release": "^19.0.5",
+    "style-loader": "^3.3.0",
+    "stylus": "^0.55.0",
+    "stylus-loader": "^7.1.0",
     "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
     "tsd": "^0.24.1",
-    "typescript": "5.1.6"
+    "typescript": "5.1.6",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.75.0",
+    "webpack-merge": "^5.8.0",
+    "@nx/next": "16.6.0"
   },
   "config": {
     "commitizen": {
