--- tools/generators/react-lib/schema.ts
@@ -1,7 +1,0 @@
-import { SupportedStyles } from '@nrwl/react/typings/style';
-
-export interface Schema  {
-  name: string;
-  style: SupportedStyles;
-  tags?: string;
-}
