+++ tools/workspace-plugin/src/generators/react-lib/schema.json
