+++ tools/workspace-plugin/src/generators/react-lib/schema.ts
@@ -1,0 +1,7 @@
+import { SupportedStyles } from '@nx/react/typings/style';
+
+export interface Schema {
+  name: string;
+  style: SupportedStyles;
+  tags?: string;
+}
