+++ tools/workspace-plugin/src/generators/ts-lib/schema.json
