+++ tools/workspace-plugin/src/generators/web-lib/schema.json
