+++ tsconfig.base.json
@@ -23,6 +23,7 @@
       "@ns3/rx-devtools": ["libs/rx-devtools/src/index.ts"],
       "@ns3/rx-state": ["libs/rx-state/src/index.ts"],
       "@ns3/ts-utils": ["libs/ts-utils/src/index.ts"],
+      "@ns3/workspace-plugin": ["tools/workspace-plugin/src/index.ts"],
       "react-demo/*": ["apps/react-demo/src/*"]
     }
   },
