--- workspace.json
@@ -1,15 +1,0 @@
-{
-  "version": 2,
-  "projects": {
-    "di": "libs/di",
-    "fetch-client": "libs/fetch-client",
-    "react-demo": "apps/react-demo",
-    "react-di": "libs/react-di",
-    "react-utils": "libs/react-utils",
-    "rx-actions": "libs/rx-actions",
-    "rx-devtools": "libs/rx-devtools",
-    "rx-state": "libs/rx-state",
-    "ts-utils": "libs/ts-utils"
-  },
-  "$schema": "./node_modules/nx/schemas/workspace-schema.json"
-}
